% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bliss_matrix.R
\name{bliss_matrix}
\alias{bliss_matrix}
\title{Calculate the additivity matrix according to Bliss' method.}
\usage{
bliss_matrix(data_init, type)
}
\arguments{
\item{data_init}{A matrix or array containing the initial data.}

\item{type}{An integer indicating the type of experiment: 2 for two-drug combinations, 3 for three-drug combinations.}
}
\value{
A matrix or array of the same dimensions as \code{data_init}, containing the calculated Bliss additivity values.
}
\description{
Calculate the additivity matrix according to Bliss' method.
}
\examples{
NULL
}
