% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_packages.R
\name{check_packages}
\alias{check_packages}
\title{Check and install required packages.}
\usage{
check_packages()
}
\description{
Check and install required packages.
}
\examples{
## Not run:
# check_packages()
## End(Not run)
}
