% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_subtable.R
\name{clean_subtable}
\alias{clean_subtable}
\title{Formatting subtable data for analysis.}
\usage{
clean_subtable(df, drug_names, type)
}
\arguments{
\item{df}{A dataframe of the subtable (repicate).}

\item{drug_names}{A list where each element is the name of the corresponding drug.}

\item{type}{An integer indicating the type of experiment: 2 for two-drug combinations, 3 for three-drug combinations.}
}
\value{
A list with two elements:
\itemize{
\item data_init: Numeric matrix/array of the response values
\item drug_doses: List of numeric vectors with the doses for each drug
}
}
\description{
Formatting subtable data for analysis.
}
\examples{
NULL
}
