% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_scientific.R
\name{convert_scientific}
\alias{convert_scientific}
\title{Convert to scientific notation.}
\usage{
convert_scientific(vect)
}
\arguments{
\item{vect}{A numeric vector}
}
\value{
A character vector with numbers in scientific notation if they have more than 6 characters.
}
\description{
Convert to scientific notation.
}
\examples{
convert_scientific(c(0.000123456, 1234, 12.3456, 1234567))
}
