\docType{data}
\name{dataC}
\alias{dataC}
\title{eDNA concentration data}
\format{A data frame containing location of eDNA sampling sites for the river Wigger (\code{dataC$ID}) and eDNA concentration values (\code{dataC$values}) (in mol m-3) for a given taxon.}
\usage{
data(dataC)
}
\description{
The dataset consists of triplicate eDNA measurements for each of the 24 sampling sites.
}
\keyword{datasets}
