\docType{data}
\name{dataCD}
\alias{dataCD}
\title{eDNA concentration and direct observation data}
\format{A data frame containing location of sampling sites for the river Wigger (\code{dataCD$ID}), respective measured values (\code{dataCD$values}) (in mol m-3 for eDNA samples; in m-2 for direct surveys) and data type (\code{dataCD$values}) (\code{"e"} for eDNA samples; \code{"d"} for direct observations) for a given taxon.}
\usage{
data(dataCD)
}
\description{
The dataset consists of triplicate eDNA measurements for each of the 24 sampling sites, plus 13 direct observation measurements.
}
\keyword{datasets}
