\docType{data}
\name{dataRead}
\alias{dataRead}
\title{eDNA read number data}
\format{A data frame containing location of eDNA sampling sites for the river Wigger (\code{dataRead$ID}) and eDNA read number values (\code{dataRead$values}) for a given taxon.}
\usage{
data(dataRead)
}
\description{
The dataset consists of triplicate eDNA measurements for each of the 24 sampling sites.
}
\keyword{datasets}
