% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{Calculate synthetic indexes according to Lehar's method.}
\usage{
index(data_init, data_bliss)
}
\arguments{
\item{data_init}{A matrix or array containing the initial data.}

\item{data_bliss}{A matrix or array of the same dimensions as \code{data_init}, containing the calculated Bliss additivity values.}
}
\value{
A list containing the Additivity Index (AI), Combination Index (CI), and Efficacy Index (EI).
}
\description{
Calculate synthetic indexes according to Lehar's method.
}
\examples{
NULL
}
