% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index.R
\name{plot_index}
\alias{plot_index}
\title{Plot the calculated indexes (AI, CI, EI) for each drug concentration.}
\usage{
plot_index(index_df)
}
\arguments{
\item{index_df}{A data frame containing the calculated indexes: Additivity Index (AI), Combination Index (CI), and Efficacy Index (EI) for each drug concentration.}
}
\value{
A patchwork plot with three bar plots for AI, CI, and EI.
}
\description{
Plot the calculated indexes (AI, CI, EI) for each drug concentration.
}
\examples{
NULL
}
