% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_title.R
\name{plot_title}
\alias{plot_title}
\title{Add a title to a plot.}
\usage{
plot_title(title)
}
\arguments{
\item{title}{The title text to be added to the plot.}
}
\value{
A grid text object representing the title, positioned at the top left of the plot.
}
\description{
Add a title to a plot.
}
\examples{
NULL
}
