% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_replicat_2drugs.R
\name{save_replicat_2drugs}
\alias{save_replicat_2drugs}
\title{Save results for replicates of 2-drug combination experiments.}
\usage{
save_replicat_2drugs(sheet_name, drug_names, global)
}
\arguments{
\item{sheet_name}{The name of the sheet being processed.}

\item{drug_names}{A list where each element is the name of the corresponding drug.}

\item{global}{A list where each element corresponds to a replicate and contains:
\itemize{
\item drug_doses: List of numeric vectors with the doses for each drug
\item data_init: Numeric matrix/array of the response values
\item data_bliss: Numeric matrix/array of the Bliss expected response value
\item index_list: List with the calculated indexes: Additivity Index (AI), Combination Index (CI), and Efficacy Index (EI)
\item heatmap_init: ComplexHeatmap object of the initial data
\item heatmap_bliss: ComplexHeatmap object of the Bliss expected data
\item heatmap_diff: ComplexHeatmap object of the difference between Bliss expected and initial data
}}
}
\value{
pdf, png and excel files saved in the output directory.
}
\description{
Save results for replicates of 2-drug combination experiments.
}
\examples{
NULL
}
