% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_drugs.R
\name{two_drugs}
\alias{two_drugs}
\title{Process data for two-drug combination experiments.}
\usage{
two_drugs(sheet_name, drug_names, sheet_data, palette = "classic")
}
\arguments{
\item{sheet_name}{The name of the sheet being processed.}

\item{drug_names}{A list where each element is the name of the corresponding drug.}

\item{sheet_data}{A dataframe of the sheet data.}

\item{palette}{Character string specifying the color palette for synergistic data. Options are "classic" (green-black-red) or "alternative" (blue-black-orange).}
}
\value{
Cf. \code{save_replicat_2drugs} function.
}
\description{
Process data for two-drug combination experiments.
}
\examples{
NULL
}
