% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap for drug combination data (data_init/data_bliss/data_diff).}
\usage{
plot_heatmap(data, drug_names, color, title = "", subtitle = "")
}
\arguments{
\item{data}{A matrix or array containing initial, additivity or synergistic data.}

\item{drug_names}{A list where each element is the name of the corresponding drug.}

\item{color}{1 for initial/additivity data, 2 for synergistic data.}

\item{title}{Character string for the title of the heatmap.}

\item{subtitle}{Character string for the subtitle of the heatmap.}
}
\value{
A ComplexHeatmap object representing the heatmap.
}
\description{
Plot heatmap for drug combination data (data_init/data_bliss/data_diff).
}
\examples{
NULL
}
