% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_EDITH.R
\name{run_EDITH}
\alias{run_EDITH}
\title{Main function to run EDITH analysis.}
\usage{
run_EDITH()
}
\value{
pdf and excel files saved in the output directory.
}
\description{
Main function to run EDITH analysis.
}
\examples{
## Not run:
# run_EDITH()
## End(Not run)
}
