% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_drugs.R
\name{three_drugs}
\alias{three_drugs}
\title{Process data for three-drug combination experiments.}
\usage{
three_drugs(sheet_name, drug_names, sheet_data)
}
\arguments{
\item{sheet_name}{The name of the sheet being processed.}

\item{drug_names}{A list where each element is the name of the corresponding drug.}

\item{sheet_data}{A dataframe of the sheet data.}
}
\value{
Cf. \code{save_perm_3drugs} function.
}
\description{
Process data for three-drug combination experiments.
}
\examples{
NULL
}
